Use double

+++ cnn/cnn/Mnist.swift
@@ -170,16 +170,16 @@ class MnistInstance
         print("----------------------------------------------")
     }
 
-    func copyImageToNNInput(aNNInputLen:Int, aNNInputArea:Int) -> [Float]
+    func copyImageToNNInput(aNNInputLen:Int, aNNInputArea:Int) -> [Double]
     {
         // 1 is white, -1 is black
-        var nnInput = [Float](count: aNNInputArea, repeatedValue: -1.0)
+        var nnInput = [Double](count: aNNInputArea, repeatedValue: -1.0)
 
         for (var ii = 0; ii < iHeight; ++ii)
         {
             for (var jj = 0; jj < iWidth; ++jj)
             {
-                let v = (Float)(iImage[ jj + iWidth*ii ])/128.0 - 1.0
+                let v = (Double)(iImage[ jj + iWidth*ii ])/128.0 - 1.0
                 //print(v)
                 nnInput[ (1 + jj) + aNNInputLen * (ii + 1) ] = v;
             }
